SystemUI: revamped battery status icon/percentage in status bar
The battery status display in the status bar has been enhanced to display
the status in four different ways:

1. Icon
2. Icon with percentage
3. Percentage only
4. Hidden

In stock Gingerbread, if you wanted to display the percentage, it would
be shown with a tiny little battery icon next to it. Now, the full sized
icon is displayed, ala Android 4.x. Also, the percentage text can now be
displayed without an icon at all, an option that was not available
previously. Finally, the percentage text has a "%" character appended to
the value, and the font size has been reduced to save space.