SystemUI: Status bar clock and battery enhancements [1/3]
- Show seconds in clock
- Show date in clock, using various styles and formats
- Show clock in various font styles
- Set custom color for clock text
- Set custom colors for battery (fill, charging bolt, percentage text)

Note that previously, when the battery style was set to portrait or landscape
icon, the charging bolt and percentage text (if set to display inside the
battery) were always transparent. With these changes, the bolt and text will
be translucent only if the system default theme is in place and no custom
colors are configured.