Fix app afterimage flashing while device unlocking
The problem is because in WMS#tryStartExitingAnimation set
win.mAnimatingExit as ture but not destroy surface when pressing
power button, so let `WindowState#mHasSurface` keeps true when
screen off. (Normally the surface should be destroyed when screen off)

This results when unlocking the screen will can see both the app
window afterimage and launcher are animating.

Check with DC#okToAnimate in tryStartExitingAnimation to ensure the
surface can be destroyed when the display is not ready to animate (e.g.
screen off).

Also, in order to cleanly clean up surfaces while cancelling recents
animation when the removal is allowed, make sure to invoke
onAnimationFinished for task's windows to finish the process of
the surface removal.

Change-Id: I0ab1faa5f0cde77fb1a337637876233bbfc707e9
CRs-Fixed: 3005096
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>